:zap: perf: try with rsc market fetch
